replaced discord.js to plain http requests. started refactor
