+++ package.json
@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@discordjs/core": "^0.5.2",
     "@headlessui/react": "^1.7.14",
     "@types/node": "18.15.11",
     "@types/react": "18.0.37",
@@ -19,6 +18,7 @@
     "postcss": "8.4.22",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "swr": "^2.1.5",
     "tailwindcss": "3.3.1",
     "typescript": "5.0.4"
   },
