+++ src/components/Feature.tsx
@@ -1,5 +1,13 @@
 import Link from "next/link";
-
+import {
+    DiscordIcon,
+    GithubIcon,
+    KnowledgeIcon,
+    TwitterIcon,
+    YoutubeIcon,
+  } from "@/components/SVGIcons";
+  import config from "@/config";
+  
 type FeatureProps = {
     title: string;
     description: string;
@@ -10,7 +18,7 @@ type FeatureProps = {
     }[]
 }
 
-export default function Feature(props: FeatureProps) {
+function Feature(props: FeatureProps) {
     return (
         <div className='max-w-sm p-6 bg-white border border-gray-200 rounded-lg shadow dark:bg-gray-800 dark:border-gray-700'>
             <h5 className="mb-2 text-2xl font-bold tracking-tight text-lightText dark:text-darkText">
@@ -32,5 +40,61 @@ export default function Feature(props: FeatureProps) {
                 ))}
             </div>
         </div>
+    )
+}
+
+export default function Features(){
+    return(
+        <div className="grid md:grid-cols-2 mt-4 gap-2">
+        <Feature
+          title={"A strong community"}
+          description={
+            "A community of blockchain researchers, developers and enthusiasts all together to unmask the secrets of blockchain technology"
+          }
+          links={[
+            {
+              icon: <TwitterIcon />,
+              text: "Follow us on twitter",
+              url: config.urls.twitter,
+            },
+            {
+              icon: <GithubIcon />,
+              text: "Star Our Repo",
+              url: config.urls.github,
+            },
+          ]}
+        />
+        <Feature
+          title={"Interviews & Podcasts"}
+          description={
+            "Interviews with people across all the web3 space and podcasts about blockchain technology"
+          }
+          links={[
+            {
+              icon: <DiscordIcon />,
+              text: "Join our discord",
+              url: config.urls.discord,
+            },
+            {
+              icon: <YoutubeIcon />,
+              text: "Listen on Youtube",
+              url: config.urls.youtube,
+            },
+          ]}
+        />
+        <Feature
+          title={"Events & Resource library"}
+          description={
+            "Time is gold so we collect almost all the info we gather into a database where we can filter out by categories"
+          }
+          links={[
+            {
+              icon: <KnowledgeIcon />,
+              text: "Get the knowledge",
+              url: config.urls.database,
+            },
+          ]}
+        />
+      </div>
     )
 }
