+++ src/components/Sponsor.tsx
@@ -1,6 +1,7 @@
+import config from "@/config";
 import Link from "next/link";
 
-export default function Sponsor(sponsor: Sponsor) {
+function Sponsor(sponsor: Sponsor) {
     return (
         <Link
             href={sponsor.url}
@@ -9,5 +10,25 @@ export default function Sponsor(sponsor: Sponsor) {
             className="p-4 grayscale dark:grayscale-0 transition duration-200 hover:grayscale-0 dark:hover:grayscale">
             {<sponsor.icon />}
         </Link>
+    )
+}
+
+export default function Sponsors() {
+    return (
+        <div className="mt-8 ">
+            <p className="text-center text-2xl dark:text-darkText  text-lightText font-bold">
+                Trusted by
+            </p>
+            <div className="grid grid-cols-3 sm:grid-cols-4 md:grid-cols-6">
+                {config.sponsors.map((sponsor) => (
+                    <Sponsor
+                        key={sponsor.url}
+                        icon={sponsor.icon}
+                        name={sponsor.name}
+                        url={sponsor.url}
+                    />
+                ))}
+            </div>
+        </div>
     )
 }
