+++ src/pages/_app.tsx
@@ -1,15 +1,11 @@
+import Layout from '@/components/Layout'
 import '@/styles/globals.css'
 import type { AppProps } from 'next/app'
-import { Roboto } from 'next/font/google'
-const roboto = Roboto({
-  weight:['400','500','700'],
-  subsets:['latin']
-})
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
-    <main className={roboto.className}>
+    <Layout>
       <Component {...pageProps} />
-    </main>
+    </Layout>
   )
 }
